refactor: make NodeApiContext independent of React Native
